+++ bgstally/ui.py
@@ -41,6 +41,9 @@ class UI:
         self.bgstally = bgstally
         self.frame = None
 
+        self.image_logo_bgstally_100 = PhotoImage(file = path.join(self.bgstally.plugin_dir, FOLDER_ASSETS, "logo_bgstally_100x67.png"))
+        self.image_logo_bgstally_16 = PhotoImage(file = path.join(self.bgstally.plugin_dir, FOLDER_ASSETS, "logo_bgstally_16x16.png"))
+        self.image_logo_bgstally_32 = PhotoImage(file = path.join(self.bgstally.plugin_dir, FOLDER_ASSETS, "logo_bgstally_32x32.png"))
         self.image_blank = PhotoImage(file = path.join(self.bgstally.plugin_dir, FOLDER_ASSETS, "blank.png"))
         self.image_button_dropdown_menu = PhotoImage(file = path.join(self.bgstally.plugin_dir, FOLDER_ASSETS, "button_dropdown_menu.png"))
         self.image_button_cmdrs = PhotoImage(file = path.join(self.bgstally.plugin_dir, FOLDER_ASSETS, "button_cmdrs.png"))
@@ -71,34 +74,32 @@ class UI:
         """
 
 
-    def get_plugin_frame(self, parent_frame:tk.Frame):
+    def get_plugin_frame(self, parent_frame: tk.Frame) -> tk.Frame:
         """
         Return a TK Frame for adding to the EDMC main window
         """
-        self.frame = tk.Frame(parent_frame)
+        self.frame: tk.Frame = tk.Frame(parent_frame)
 
-        current_row = 0
-        tk.Label(self.frame, text="BGS Tally").grid(row=current_row, column=0, sticky=tk.W)
+        current_row: int = 0
+        tk.Label(self.frame, image=self.image_logo_bgstally_100).grid(row=current_row, column=0, rowspan=3, sticky=tk.W)
         self.label_version = HyperlinkLabel(self.frame, text=f"v{str(self.bgstally.version)}", background=nb.Label().cget('background'), url=URL_LATEST_RELEASE, underline=True)
         self.label_version.grid(row=current_row, column=1, columnspan=3 if self.bgstally.capi_fleetcarrier_available() else 2, sticky=tk.W)
+        current_row += 1
+        tk.Label(self.frame, text="BGS Tally Status:" + " " + self.bgstally.state.Status.get()).grid(row=current_row, column=1, sticky=tk.W)
+        current_row += 1
+        self.label_tick = tk.Label(self.frame, text="Last BGS Tick:" + " " + self.bgstally.tick.get_formatted())
+        self.label_tick.grid(row=current_row, column=1, sticky=tk.W)
         current_row += 1
         self.button_latest_tick: tk.Button = tk.Button(self.frame, text="Latest BGS Tally", height=SIZE_BUTTON_PIXELS-2, image=self.image_blank, compound=tk.RIGHT, command=partial(self._show_activity_window, self.bgstally.activity_manager.get_current_activity()))
         self.button_latest_tick.grid(row=current_row, column=0, padx=3)
         self.button_previous_ticks: tk.Button = tk.Button(self.frame, text="Previous BGS Tallies ", height=SIZE_BUTTON_PIXELS-2, image=self.image_button_dropdown_menu, compound=tk.RIGHT, command=self._previous_ticks_popup)
-        self.button_previous_ticks.grid(row=current_row, column=1, padx=3)
+        self.button_previous_ticks.grid(row=current_row, column=1, padx=3, sticky=tk.W)
         tk.Button(self.frame, image=self.image_button_cmdrs, height=SIZE_BUTTON_PIXELS, width=SIZE_BUTTON_PIXELS, command=self._show_cmdr_list_window).grid(row=current_row, column=2, padx=3)
         if self.bgstally.capi_fleetcarrier_available():
             self.button_carrier: tk.Button = tk.Button(self.frame, image=self.image_button_carrier, state=('normal' if self.bgstally.fleet_carrier.available() else 'disabled'), height=SIZE_BUTTON_PIXELS, width=SIZE_BUTTON_PIXELS, command=self._show_fc_window)
             self.button_carrier.grid(row=current_row, column=3, padx=3)
         else:
             self.button_carrier: tk.Button = None
-        current_row += 1
-        tk.Label(self.frame, text="BGS Tally Status:").grid(row=current_row, column=0, sticky=tk.W)
-        tk.Label(self.frame, textvariable=self.bgstally.state.Status).grid(row=current_row, column=1, sticky=tk.W)
-        current_row += 1
-        tk.Label(self.frame, text="Last BGS Tick:").grid(row=current_row, column=0, sticky=tk.W)
-        self.label_tick: tk.Label = tk.Label(self.frame, text=self.bgstally.tick.get_formatted())
-        self.label_tick.grid(row=current_row, column=1, sticky=tk.W)
         current_row += 1
 
         return self.frame
@@ -115,7 +116,7 @@ class UI:
         else:
             self.label_version.configure(text=f"v{str(self.bgstally.version)}", url=URL_LATEST_RELEASE, foreground='blue')
 
-        self.label_tick.config(text=self.bgstally.tick.get_formatted())
+        self.label_tick.config(text="Last BGS Tick:" + " " + self.bgstally.tick.get_formatted())
         self.button_latest_tick.config(command=partial(self._show_activity_window, self.bgstally.activity_manager.get_current_activity()))
         if self.button_carrier is not None:
             self.button_carrier.config(state=('normal' if self.bgstally.fleet_carrier.available() else 'disabled'))